+++ apps/common/lib/lexical/ast.ex
@@ -384,91 +384,6 @@ defmodule Lexical.Ast do
     end
   end
 
-  @doc """
-  Expands an alias at the given position in the context of a document
-  analysis.
-
-  When we refer to a module, it's usually a short name, often aliased or
-  in a nested module. This function finds the full name of the module at
-  a cursor position.
-
-  For example, if we have:
-
-      defmodule Project do
-        defmodule Issue do
-          defstruct [:message]
-        end
-
-        def message(%Issue{|} = issue) do # cursor marked as `|`
-        end
-      end
-
-  We could get the expansion for the `Issue` alias at the cursor position
-  like so:
-
-      iex> Ast.expand_alias([:Issue], analysis, position)
-      {:ok, Project.Issue}
-
-  Another example:
-
-      defmodule Project do
-        defmodule Issue do
-          defstruct [:message]
-        end
-      end
-
-      defmodule MyModule do
-        alias Project, as: MyProject
-
-        def message(%MyProject.Issue{|} = issue) do
-        end
-      end
-
-  This would yield the same result:
-
-      iex> Ast.expand_alias([:MyProject, :Issue], analysis, position)
-      {:ok, Project.Issue}
-
-  If no aliases are present at the given position, no expansion occurs:
-
-      iex> Ast.expand_alias([:Some, :Other, :Module], analysis, position)
-      {:ok, Some.Other.Module}
-
-  """
-  @spec expand_alias(
-          alias_segments() | module(),
-          Analysis.t(),
-          Position.t() | {Position.line(), Position.character()}
-        ) ::
-          {:ok, module()} | :error
-  def expand_alias([_ | _] = segments, %Analysis{} = analysis, %Position{} = position) do
-    with %Analysis{valid?: true} = analysis <- reanalyze_to(analysis, position),
-         aliases <- Analysis.aliases_at(analysis, position),
-         {:ok, resolved} <- resolve_alias(segments, aliases) do
-      {:ok, Module.concat(resolved)}
-    else
-      _ ->
-        if Enum.all?(segments, &is_atom/1) do
-          {:ok, Module.concat(segments)}
-        else
-          :error
-        end
-    end
-  end
-
-  def expand_alias(module, %Analysis{} = analysis, %Position{} = position)
-      when is_atom(module) and not is_nil(module) do
-    module
-    |> Module.split()
-    |> Enum.map(&String.to_atom/1)
-    |> expand_alias(analysis, position)
-  end
-
-  def expand_alias(empty, _, _) when empty in [nil, []] do
-    Logger.warning("nothing to expand (expand_alias was passed #{inspect(empty)})")
-    :error
-  end
-
   # Expands aliases given the rules in the special form
   # https://hexdocs.pm/elixir/1.13.4/Kernel.SpecialForms.html#__aliases__/1
   def reify_alias(current_module, [:"Elixir" | _] = reified) do
@@ -491,18 +406,6 @@ defmodule Lexical.Ast do
   end
 
   # private
-
-  defp resolve_alias([first | _] = segments, aliases_mapping) when is_tuple(first) do
-    with {:ok, current_module} <- Map.fetch(aliases_mapping, :__MODULE__) do
-      {:ok, reify_alias(current_module, segments)}
-    end
-  end
-
-  defp resolve_alias([first | rest], aliases_mapping) do
-    with {:ok, resolved} <- Map.fetch(aliases_mapping, first) do
-      {:ok, [resolved | rest]}
-    end
-  end
 
   defp do_string_to_quoted(string) when is_binary(string) do
     Code.string_to_quoted(string,
