+++ apps/common/lib/lexical/ast/analysis/analyzer.ex
@@ -1,386 +1,18 @@
 defmodule Lexical.Ast.Analysis.Analyzer do
   @moduledoc false
 
-  alias __MODULE__
+  alias Lexical.Ast.Analysis.Analyzer.Alias
+  alias Lexical.Ast.Analysis.Analyzer.Import
+  alias Lexical.Ast.Analysis.Analyzer.Scope
+  alias Lexical.Ast.Analysis.Analyzer.State
   alias Lexical.Document
   alias Lexical.Document.Position
-  alias Lexical.Document.Range
-  alias Lexical.Module.Loader
   alias Sourceror.Zipper
 
   @scope_id :_scope_id
 
   @block_keywords [:do, :else, :rescue, :catch, :after]
   @clauses [:->]
-
-  defmodule Alias do
-    defstruct [:module, :as, :line]
-
-    @type t :: %Alias{}
-
-    def new(module, as, line) when is_list(module) and is_atom(as) and line > 0 do
-      %Alias{module: module, as: as, line: line}
-    end
-
-    def to_module(%Alias{} = alias) do
-      Module.concat(alias.module)
-    end
-  end
-
-  defmodule Import do
-    alias Lexical.Ast.Analysis.Analyzer.Scope
-    alias Lexical.ProcessCache
-
-    defstruct module: nil, selector: :all, line: nil
-    @type function_name :: atom()
-    @type function_arity :: {function_name(), arity()}
-    @type selector ::
-            :functions | :macros | [only: [function_arity()]] | [except: [function_arity()]]
-    @type t :: %{
-            module: module(),
-            selector: selector(),
-            line: non_neg_integer()
-          }
-    def new(module, line) do
-      %__MODULE__{module: module, line: line}
-    end
-
-    def new(module, selector, line) do
-      %__MODULE__{module: module, selector: expand_selector(selector), line: line}
-    end
-
-    def apply_to_scope(%__MODULE__{} = import, current_scope, %{} = current_imports) do
-      import_module = Scope.resolve_alias_at(current_scope, import.module, import.line)
-
-      functions = mfas_for(import_module, :functions)
-      macros = mfas_for(import_module, :macros)
-
-      case import.selector do
-        :all ->
-          Map.put(current_imports, import_module, functions ++ macros)
-
-        [only: :functions] ->
-          Map.put(current_imports, import_module, functions)
-
-        [only: :macros] ->
-          Map.put(current_imports, import_module, macros)
-
-        [only: functions_to_import] ->
-          functions_to_import = function_and_arity_to_mfa(import_module, functions_to_import)
-          Map.put(current_imports, import_module, functions_to_import)
-
-        [except: functions_to_except] ->
-          # This one is a little tricky. Imports using except have two cases.
-          # In the first case, if the module hasn't been previously imported, we
-          # collect all the functions in the current module and remove the ones in the
-          # except clause.
-          # If the module has been previously imported, we just remove the functions from
-          # the except clause from those that have been previously imported.
-          # See: https://hexdocs.pm/elixir/1.13.0/Kernel.SpecialForms.html#import/2-selector
-
-          functions_to_except = function_and_arity_to_mfa(import_module, functions_to_except)
-
-          if already_imported?(current_imports, import_module) do
-            Map.update!(current_imports, import_module, fn old_imports ->
-              old_imports -- functions_to_except
-            end)
-          else
-            to_import = (functions ++ macros) -- functions_to_except
-            Map.put(current_imports, import_module, to_import)
-          end
-      end
-    end
-
-    defp expand_selector(selectors) do
-      selectors =
-        Enum.reduce(selectors, [], fn
-          {{:__block__, _, [type]}, {:__block__, _, [selector]}}, acc
-          when type in [:only, :except] ->
-            expanded =
-              case selector do
-                :functions ->
-                  :functions
-
-                :macros ->
-                  :macros
-
-                keyword when is_list(keyword) ->
-                  Enum.map(keyword, fn
-                    {{:__block__, _, [function_name]}, {:__block__, _, [arity]}} ->
-                      {function_name, arity}
-                  end)
-              end
-
-            [{type, expanded} | acc]
-
-          _, acc ->
-            acc
-        end)
-
-      if selectors == [] do
-        :all
-      else
-        selectors
-      end
-    end
-
-    defp already_imported?(%{} = current_imports, imported_module) do
-      case current_imports do
-        %{^imported_module => [_ | _]} -> true
-        _ -> false
-      end
-    end
-
-    defp function_and_arity_to_mfa(current_module, fa_list) when is_list(fa_list) do
-      Enum.map(fa_list, fn {function, arity} -> {current_module, function, arity} end)
-    end
-
-    defp mfas_for(current_module, type) do
-      if Loader.ensure_loaded?(current_module) do
-        fa_list = function_and_arities_for_module(current_module, type)
-
-        function_and_arity_to_mfa(current_module, fa_list)
-      else
-        []
-      end
-    end
-
-    defp function_and_arities_for_module(module, type) do
-      ProcessCache.trans({module, :info, type}, fn ->
-        type
-        |> module.__info__()
-        |> Enum.reject(fn {name, _arity} ->
-          name |> Atom.to_string() |> String.starts_with?("_")
-        end)
-      end)
-    end
-  end
-
-  defmodule Scope do
-    defstruct [
-      :id,
-      :range,
-      module: [],
-      aliases: [],
-      imports: []
-    ]
-
-    @kernel_imports [
-      Import.new([:Kernel], 1),
-      Import.new([:Kernel, :SpecialForms], 1)
-    ]
-
-    @type import_mfa :: {module(), atom(), non_neg_integer()}
-    @type scope_position :: Position.t() | :end
-    @blank_doc Document.new("file:///", "", 1)
-
-    @type t :: %__MODULE__{
-            id: any(),
-            range: Range.t(),
-            module: [atom()],
-            aliases: [Alias.t()],
-            imports: [import_mfa()]
-          }
-
-    def new(%__MODULE__{} = parent_scope, id, %Range{} = range, module \\ []) do
-      %Scope{
-        id: id,
-        aliases: parent_scope.aliases,
-        imports: parent_scope.imports,
-        module: module,
-        range: range
-      }
-    end
-
-    def global(%Range{} = range) do
-      %Scope{id: :global, range: range}
-    end
-
-    @spec imports(t(), scope_position()) :: [import_mfa()]
-    def imports(%__MODULE__{} = scope, position \\ :end) do
-      scope
-      |> import_map(position)
-      |> Map.values()
-      |> List.flatten()
-    end
-
-    def import_map(%__MODULE__{} = scope, position \\ :end) do
-      end_line = end_line(scope, position)
-
-      (@kernel_imports ++ scope.imports)
-      # sorting by line ensures that imports on later lines
-      # override imports on earlier lines
-      |> Enum.sort_by(& &1.line)
-      |> Enum.take_while(&(&1.line <= end_line))
-      |> Enum.reduce(%{}, fn %Import{} = import, current_imports ->
-        Import.apply_to_scope(import, scope, current_imports)
-      end)
-    end
-
-    @spec alias_map(Scope.t(), scope_position()) :: %{module() => Scope.t()}
-    def alias_map(%Scope{} = scope, position \\ :end) do
-      end_line = end_line(scope, position)
-
-      scope.aliases
-      # sorting by line ensures that aliases on later lines
-      # override aliases on earlier lines
-      |> Enum.sort_by(& &1.line)
-      |> Enum.take_while(&(&1.line <= end_line))
-      |> Map.new(&{&1.as, &1})
-    end
-
-    def resolve_alias_at(%__MODULE__{} = scope, module, line) do
-      position = Position.new(@blank_doc, line, 1)
-      aliases = alias_map(scope, position)
-
-      case module do
-        [{:__MODULE__, _, _} | suffix] ->
-          current_module =
-            aliases
-            |> Map.get(:__MODULE__)
-            |> Alias.to_module()
-
-          Module.concat([current_module | suffix])
-
-        [prefix | suffix] ->
-          case aliases do
-            %{^prefix => _} ->
-              current_module =
-                aliases
-                |> Map.get(prefix)
-                |> Alias.to_module()
-
-              Module.concat([current_module | suffix])
-
-            _ ->
-              Module.concat(module)
-          end
-      end
-    end
-
-    def fetch_alias_with_prefix(%__MODULE__{} = scope, prefix) do
-      case Enum.find(scope.aliases, fn %Alias{} = alias -> alias.as == prefix end) do
-        %Alias{} = existing -> {:ok, existing}
-        _ -> :error
-      end
-    end
-
-    def empty?(%Scope{aliases: [], imports: []}), do: true
-    def empty?(%Scope{}), do: false
-
-    defp end_line(%__MODULE__{} = scope, :end), do: scope.range.end.line
-    defp end_line(_, %Position{} = position), do: position.line
-  end
-
-  defmodule State do
-    defstruct [:document, scopes: [], visited: %{}]
-
-    def new(%Document{} = document) do
-      state = %State{document: document}
-
-      scope =
-        document
-        |> global_range()
-        |> Scope.global()
-
-      push_scope(state, scope)
-    end
-
-    def current_scope(%State{scopes: [scope | _]}), do: scope
-
-    def current_module(%State{} = state) do
-      current_scope(state).module
-    end
-
-    def push_scope(%State{} = state, %Scope{} = scope) do
-      Map.update!(state, :scopes, &[scope | &1])
-    end
-
-    def push_scope(%State{} = state, id, %Range{} = range, module) when is_list(module) do
-      scope =
-        state
-        |> current_scope()
-        |> Scope.new(id, range, module)
-
-      push_scope(state, scope)
-    end
-
-    def push_scope_for(%State{} = state, quoted, %Range{} = range, module) do
-      module = module || current_module(state)
-      id = Analyzer.scope_id(quoted)
-      push_scope(state, id, range, module)
-    end
-
-    def push_scope_for(%State{} = state, quoted, module) do
-      range = get_range(quoted, state.document)
-      push_scope_for(state, quoted, range, module)
-    end
-
-    def maybe_push_scope_for(%State{} = state, quoted) do
-      case get_range(quoted, state.document) do
-        %Range{} = range ->
-          push_scope_for(state, quoted, range, nil)
-
-        nil ->
-          state
-      end
-    end
-
-    def pop_scope(%State{scopes: [scope | rest]} = state) do
-      %State{state | scopes: rest, visited: Map.put(state.visited, scope.id, scope)}
-    end
-
-    def push_alias(%State{} = state, %Alias{} = alias) do
-      update_current_scope(state, fn %Scope{} = scope ->
-        [prefix | rest] = alias.module
-
-        alias =
-          case Scope.fetch_alias_with_prefix(scope, prefix) do
-            {:ok, %Alias{} = existing_alias} ->
-              %Alias{alias | module: existing_alias.module ++ rest}
-
-            :error ->
-              alias
-          end
-
-        Map.update!(scope, :aliases, &[alias | &1])
-      end)
-    end
-
-    def push_import(%State{} = state, %Import{} = import) do
-      update_current_scope(state, fn %Scope{} = scope ->
-        Map.update!(scope, :imports, &[import | &1])
-      end)
-    end
-
-    defp update_current_scope(%State{} = state, fun) do
-      update_in(state, [Access.key(:scopes), Access.at!(0)], fn %Scope{} = scope ->
-        fun.(scope)
-      end)
-    end
-
-    defp get_range(quoted, %Document{} = document) do
-      case Sourceror.get_range(quoted) do
-        %{start: start_pos, end: end_pos} ->
-          Range.new(
-            Position.new(document, start_pos[:line], start_pos[:column]),
-            Position.new(document, end_pos[:line], end_pos[:column])
-          )
-
-        nil ->
-          nil
-      end
-    end
-
-    defp global_range(%Document{} = document) do
-      num_lines = Document.size(document)
-
-      Range.new(
-        Position.new(document, 1, 1),
-        Position.new(document, num_lines + 1, 1)
-      )
-    end
-  end
 
   @doc """
   Traverses an AST, returning a list of scopes.
@@ -453,9 +85,15 @@ defmodule Lexical.Ast.Analysis.Analyzer do
         # we go up twice to get to the real parent because ast pairs
         # are always in a list
         %Zipper{node: parent} = zipper |> Zipper.up() |> Zipper.up()
-        parent_end = Sourceror.get_range(parent).end
-        new_end = Position.new(document, parent_end[:line], parent_end[:column])
-        put_in(scope.range.end, new_end)
+
+        case Sourceror.get_range(parent) do
+          %{end: parent_end} ->
+            new_end = Position.new(document, parent_end[:line], parent_end[:column])
+            put_in(scope.range.end, new_end)
+
+          _ ->
+            scope
+        end
     end
   end
