+++ apps/common/mix.exs
@@ -18,8 +18,7 @@ defmodule Common.MixProject do
 
   def application do
     [
-      extra_applications: [:logger],
-      mod: {Lexical.Common.Application, []}
+      extra_applications: [:logger]
     ]
   end
