+++ apps/common/test/test_helper.exs
@@ -1,2 +1,1 @@
-Application.ensure_all_started(:common)
 ExUnit.start()
