+++ apps/remote_control/lib/lexical/remote_control/application.ex
@@ -7,12 +7,16 @@ defmodule Lexical.RemoteControl.Application do
   alias Lexical.RemoteControl
 
   use Application
+  require Logger
 
   @impl true
   def start(_type, _args) do
+    Logger.info("Remote control: indexing? #{Features.indexing_enabled?()}")
+
     children =
       if RemoteControl.project_node?() do
         [
+          RemoteControl.Module.Loader,
           {RemoteControl.Dispatch, progress: true},
           RemoteControl.ModuleMappings,
           RemoteControl.Build,
@@ -26,7 +30,6 @@ defmodule Lexical.RemoteControl.Application do
       end
 
     children = Enum.reject(children, &is_nil/1)
-
     # See https://hexdocs.pm/elixir/Supervisor.html
     # for other strategies and supported options
     opts = [strategy: :one_for_one, name: Lexical.RemoteControl.Supervisor]
