+++ apps/remote_control/lib/lexical/remote_control/search/indexer/extractors/function_definition.ex
@@ -1,10 +1,10 @@
 defmodule Lexical.RemoteControl.Search.Indexer.Extractors.FunctionDefinition do
-  alias Lexical.Ast
   alias Lexical.Ast.Analysis
   alias Lexical.Document
   alias Lexical.Document.Position
   alias Lexical.Document.Range
   alias Lexical.Formats
+  alias Lexical.RemoteControl
   alias Lexical.RemoteControl.Search.Indexer.Entry
   alias Lexical.RemoteControl.Search.Indexer.Metadata
   alias Lexical.RemoteControl.Search.Indexer.Source.Block
@@ -15,7 +15,9 @@ defmodule Lexical.RemoteControl.Search.Indexer.Extractors.FunctionDefinition do
   def extract({definition, metadata, [{fn_name, _, args}, _body]} = ast, %Reducer{} = reducer)
       when is_atom(fn_name) and definition in @function_definitions do
     range = get_definition_range(reducer.analysis, metadata)
-    {:ok, module} = Ast.expand_alias([:__MODULE__], reducer.analysis, range.start)
+
+    {:ok, module} =
+      RemoteControl.Analyzer.expand_alias([:__MODULE__], reducer.analysis, range.start)
 
     arity =
       case args do
