+++ apps/remote_control/test/test_helper.exs
@@ -5,5 +5,5 @@ with :nonode@nohost <- Node.self() do
   {:ok, _pid} = :net_kernel.start([:"testing-#{random_number}@127.0.0.1"])
 end
 
-Application.ensure_all_started(:common)
+Lexical.RemoteControl.Module.Loader.start_link(nil)
 ExUnit.start(exclude: [:skip])
