+++ apps/server/test/test_helper.exs
@@ -1,3 +1,2 @@
-Application.ensure_all_started(:common)
 ExUnit.configure(timeout: :infinity)
 ExUnit.start()
